Solved the navigation issue.
The reason of view automatically got popped out in second run is because the view is popped in the second time within the navigation path. Seems that this path works rather like iterating in depth in a single line within a tree. The idea is to use `removeLast` to pop the view back to its root (`ApplicationListView`) however the pop index should be calculated (1 here).

Pathmanager is now passed explicitly into every views as a binding, just for testing and ensuring things. This may need to be changed in the future. There is an `update navigationauthority` issue.

Another issue is related to the creation of job description objects, as the ctor way may work in some times but will crash in others. Consequently, relationship properties are marked as nullable and intialized first to nil then once the object get created, relationships are appended.