Main JSON decode issue should be fixed so far
The issue was the navigation direction of one-to-many relationship from another side. Since the current data model could be represented by a tree from the `Company` side, the JSON decode order was changed in this way. Now importing data from JSON is working without error.

There is another issue that some entries will be omitted, like for a sample data of 5 applications, only 3 will be imported. This issue may be related to the enum `ApplicationStatus` naming convention since both omitted entries have underscores in status names.