Fixed time window for updates; also fixed career type duplicate issue
Seems that the `context.fetch` method doesn't work during JSON decode (`required init`) phases (it will fetch nothing but a nil), the de-duplication was handled once the JSON decode is finished.

The process is to trigger every JD with duplicated careers to their first occurrences, and then remove these duplicata.